[Fix#49] 예산 설정 타입 버그 수정
int -> long 으로 타입 변경

+++ data/src/main/java/com/aos/data/entity/request/book/PostBooksInfoBudgetBody.kt
@@ -5,6 +5,6 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class PostBooksInfoBudgetBody (
     val bookKey : String,
-    val budget : Int,
+    val budget : Long,
     val date : String
 )
